requirements.txt
fastapi
uvicorn[standard]
black
flake8
web3>=6.0.0
python-dotenv>=0.19.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.0.0
base58>=2.1.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
python-multipart>=0.0.6